Split bs cli into multiple files
