+++ cmd/bs/main.go
@@ -1,22 +1,11 @@
 package main
 
 import (
-	"context"
-	"fmt"
-	"github.com/anthdm/bs/db"
 	"github.com/anthdm/bs/migrations"
-	"github.com/uptrace/bun/migrate"
 	"github.com/urfave/cli/v2"
 	"log"
 	"os"
-	"strings"
 )
-
-func newMigrator(migrations *migrate.Migrations) *migrate.Migrator {
-	db.Init()
-
-	return migrate.NewMigrator(db.Bun, migrations)
-}
 
 func main() {
 	app := &cli.App{
@@ -27,161 +16,5 @@ func main() {
 	}
 	if err := app.Run(os.Args); err != nil {
 		log.Fatal(err)
-	}
-}
-
-func newDBCommand(migrations *migrate.Migrations) *cli.Command {
-	return &cli.Command{
-		Name:  "db",
-		Usage: "manage database migrations",
-		Subcommands: []*cli.Command{
-			{
-				Name:  "init",
-				Usage: "create migration tables",
-				Action: func(c *cli.Context) error {
-					ctx := context.Background()
-					migrator := newMigrator(migrations)
-					return migrator.Init(ctx)
-				},
-			},
-			{
-				Name:  "migrate",
-				Usage: "migrate database",
-				Action: func(c *cli.Context) error {
-					ctx := context.Background()
-					migrator := newMigrator(migrations)
-
-					group, err := migrator.Migrate(ctx)
-					if err != nil {
-						return err
-					}
-
-					if group.ID == 0 {
-						fmt.Printf("there are no new migrations to run\n")
-						return nil
-					}
-
-					fmt.Printf("migrated to %s\n", group)
-					return nil
-				},
-			},
-			{
-				Name:  "rollback",
-				Usage: "rollback the last migration group",
-				Action: func(c *cli.Context) error {
-					ctx := context.Background()
-					migrator := newMigrator(migrations)
-
-					group, err := migrator.Rollback(ctx)
-					if err != nil {
-						return err
-					}
-
-					if group.ID == 0 {
-						fmt.Printf("there are no groups to roll back\n")
-						return nil
-					}
-
-					fmt.Printf("rolled back %s\n", group)
-					return nil
-				},
-			},
-			{
-				Name:  "lock",
-				Usage: "lock migrations",
-				Action: func(c *cli.Context) error {
-					ctx := context.Background()
-					migrator := migrate.NewMigrator(db.Bun, migrations)
-
-					return migrator.Lock(ctx)
-				},
-			},
-			{
-				Name:  "unlock",
-				Usage: "unlock migrations",
-				Action: func(c *cli.Context) error {
-					ctx := context.Background()
-					migrator := migrate.NewMigrator(db.Bun, migrations)
-
-					return migrator.Unlock(ctx)
-				},
-			},
-			{
-				Name:  "create_go",
-				Usage: "create Go migration",
-				Action: func(c *cli.Context) error {
-					ctx := context.Background()
-					migrator := migrate.NewMigrator(db.Bun, migrations)
-
-					name := strings.Join(c.Args().Slice(), "_")
-					mf, err := migrator.CreateGoMigration(ctx, name)
-					if err != nil {
-						return err
-					}
-					fmt.Printf("created migration %s (%s)\n", mf.Name, mf.Path)
-
-					return nil
-				},
-			},
-			{
-				Name:  "create_sql",
-				Usage: "create up and down SQL migrations",
-				Action: func(c *cli.Context) error {
-					ctx := context.Background()
-					migrator := migrate.NewMigrator(db.Bun, migrations)
-
-					name := strings.Join(c.Args().Slice(), "_")
-					files, err := migrator.CreateSQLMigrations(ctx, name)
-					if err != nil {
-						return err
-					}
-
-					for _, mf := range files {
-						fmt.Printf("created migration %s (%s)\n", mf.Name, mf.Path)
-					}
-
-					return nil
-				},
-			},
-			{
-				Name:  "status",
-				Usage: "print migrations status",
-				Action: func(c *cli.Context) error {
-					ctx := context.Background()
-					migrator := migrate.NewMigrator(db.Bun, migrations)
-
-					ms, err := migrator.MigrationsWithStatus(ctx)
-					if err != nil {
-						return err
-					}
-					fmt.Printf("migrations: %s\n", ms)
-					fmt.Printf("unapplied migrations: %s\n", ms.Unapplied())
-					fmt.Printf("last migration group: %s\n", ms.LastGroup())
-
-					return nil
-				},
-			},
-			{
-				Name:  "mark_applied",
-				Usage: "mark migrations as applied without actually running them",
-				Action: func(c *cli.Context) error {
-					ctx := context.Background()
-					migrator := migrate.NewMigrator(db.Bun, migrations)
-
-					group, err := migrator.Migrate(ctx, migrate.WithNopMigration())
-					if err != nil {
-						return err
-					}
-
-					if group.ID == 0 {
-						fmt.Printf("there are no new migrations to mark as applied\n")
-						return nil
-					}
-
-					fmt.Printf("marked as applied %s\n", group)
-					return nil
-				},
-			},
-		},
 	}
 }
